+++ src/drivers/gfx.h
@@ -37,8 +37,8 @@ void gfx_set_buf(unsigned char buf);
 
 void gfx_fill(color32_t color);
 void gfx_draw_filled_rect(p2d_t pos, p2d_t size, color32_t c);
-void gfx_draw_hor_line(p2d_t pos, uint16_t w, color32_t c);
-void gfx_draw_vert_line(p2d_t pos, uint16_t h, color32_t c);
+void gfx_draw_hor_line(p2d_t pos, uint64_t w, color32_t c);
+void gfx_draw_vert_line(p2d_t pos, uint64_t h, color32_t c);
 void gfx_draw_rect(p2d_t pos, p2d_t size, color32_t c);
 void gfx_draw_xbm(p2d_t position, uint8_t* xbm_ptr, p2d_t xbm_size, color32_t color_h, color32_t color_l);
 void gfx_draw_raw(p2d_t position, uint8_t* raw_ptr, p2d_t raw_size);
